+++ application/views/dashboard/medicine/view.php
@@ -13,7 +13,7 @@ $this->load->view('dashboard/structure/_leftmenu', array('heading' => $heading))
 	<p>
 	<span class="back"><a href="<?= base_url()?>medicine">All medicine</a></span>
   	<span class="edit"><a href="<?= base_url()?>medicine/edit/<?=$medicine->id?>">Edit details</a></span>
-  	<span class="delete"><a href="<?= base_url()?>medicine/delete/<?=$medicine->id?>">Delete this medicine</a></span>
+  	<span class="delete"><a onclick="return confirm('Are you sure you wish to delete?');" href="<?= base_url()?>medicine/delete/<?=$medicine->id?>">Delete this medicine</a></span>
   	<span class="eye"><a href="<?= base_url()?>medicine/index/<?=$medicine->name?>">Same name medicines</a></span>
 	</p>
